fix: resolve enum value in other file
