+++ pilota-build/test_data/thrift/default_value.thrift
@@ -1,5 +1,11 @@
+enum B {
+    Read = 1;
+    Write = 2;
+}
+
 struct A {
     1: required string faststr = "hello world",
     2: required string string = "test"(pilota.rust_type = "string"),
     3: optional bool a = false,
+    4: optional B test_b = B.Read,
 }
